Refactored: Datadog::Logger --> Datadog::Core::Logger
