+++ lib/ddtrace/tracer.rb
@@ -9,7 +9,7 @@ require 'ddtrace/context_provider'
 require 'ddtrace/context'
 require 'ddtrace/correlation'
 require 'ddtrace/event'
-require 'ddtrace/logger'
+require 'datadog/core/logger'
 require 'ddtrace/sampler'
 require 'ddtrace/sampling'
 require 'ddtrace/span_operation'
