+++ spec/ddtrace/tracer_spec.rb
@@ -496,7 +496,7 @@ RSpec.describe Datadog::Tracer do
             let(:not_a_proc_block) { 'not a proc' }
 
             it 'fallbacks to default error handler and log a debug message' do
-              expect_any_instance_of(Datadog::Logger).to receive(:debug).at_least(:once)
+              expect_any_instance_of(Datadog::Core::Logger).to receive(:debug).at_least(:once)
               expect do
                 tracer.trace(name, on_error: not_a_proc_block, &block)
               end.to raise_error(error)
